Remove v1 folder structure (#467)
* Change Admin to Config
* Move tags definitions to the app level
* Remove "v1" for every tag as we don't have any other versions of the API. If one day we do, we can add v1 back.
* Remove v1 folder
